Extract select option normalization into a utility
